backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
pyodbc
pydantic
python-jose[cryptography]
bcrypt
python-multipart
